added more auth stuff
