+++ backend/routes/users.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const {registerUser, loginUser} = require('../controllers/userController');
+
 const router = express.Router();
 
 router.post('/login', loginUser); // login
