+++ frontend/src/Hooks/useUserContext.js
@@ -3,7 +3,7 @@ import { useContext } from "react";
 
 
 export const useUserContext = () => {
-    const context = useContext(UserContext)
+    const context = useContext(userContext)
 
     if(!context){
         throw Error('useUserContext must be use inside a userContextProvider')
